DAC6-4134 | separate classes into their own respective files
