+++ app/models/FinancialInstitutions/CreateFIDetails.scala
@@ -14,19 +14,21 @@
  * limitations under the License.
  */
 
-package models.response
+package models.FinancialInstitutions
 
-import play.api.libs.json.{Json, OFormat}
+import play.api.libs.json._
 
-case class ErrorDetail(
-  timestamp: String,
-  correlationId: String,
-  errorCode: Option[String] = None,
-  errorMessage: Option[String] = None,
-  source: Option[String] = None,
-  sourceFaultDetail: Option[SourceFaultDetail] = None
+final case class CreateFIDetails(
+  FIName: String,
+  SubscriptionID: String,
+  TINDetails: Option[Seq[TINDetails]],
+  GIIN: Option[String],
+  IsFIUser: Boolean,
+  AddressDetails: AddressDetails,
+  PrimaryContactDetails: Option[ContactDetails],
+  SecondaryContactDetails: Option[ContactDetails]
 )
 
-object ErrorDetail {
-  implicit val errorDetailReads: OFormat[ErrorDetail] = Json.format[ErrorDetail]
+object CreateFIDetails {
+  implicit val format: OFormat[CreateFIDetails] = Json.format[CreateFIDetails]
 }
