+++ app/models/FinancialInstitutions/FIDetails.scala
@@ -16,9 +16,7 @@
 
 package models.FinancialInstitutions
 
-import models.Address
 import play.api.libs.json._
-import utils.CountryListFactory
 
 final case class FIDetails(
   FIID: String,
@@ -34,76 +32,5 @@ final case class FIDetails(
 
 object FIDetails {
   implicit val format: OFormat[FIDetails] = Json.format[FIDetails]
-}
-
-final case class RemoveFIDetail(
-  SubscriptionID: String,
-  FIID: String
-)
-
-object RemoveFIDetail {
-  implicit val format: OFormat[RemoveFIDetail] = Json.format[RemoveFIDetail]
-}
-
-object AddressDetails {
-  implicit val format: OFormat[AddressDetails] = Json.format[AddressDetails]
-
-  implicit class AddressDetailsExtension(addressDetails: AddressDetails) {
-
-    def toAddress(factory: CountryListFactory): Option[Address] =
-      for {
-        countryCode <- addressDetails.CountryCode
-        country     <- factory.findCountryWithCode(countryCode)
-      } yield Address(
-        addressLine1 = addressDetails.AddressLine1,
-        addressLine2 = addressDetails.AddressLine2,
-        addressLine3 = addressDetails.AddressLine3,
-        addressLine4 = addressDetails.AddressLine4,
-        postCode = addressDetails.PostalCode,
-        country = country
-      )
-
-  }
-
-}
-
-final case class AddressDetails(
-  AddressLine1: String,
-  AddressLine2: Option[String],
-  AddressLine3: Option[String],
-  AddressLine4: Option[String],
-  CountryCode: Option[String],
-  PostalCode: Option[String]
-)
-
-final case class ContactDetails(ContactName: String, EmailAddress: String, PhoneNumber: Option[String])
-
-object ContactDetails {
-  implicit val format: OFormat[ContactDetails] = Json.format[ContactDetails]
-}
-
-final case class CreateFIDetails(
-  FIName: String,
-  SubscriptionID: String,
-  TINDetails: Option[Seq[TINDetails]],
-  GIIN: Option[String],
-  IsFIUser: Boolean,
-  AddressDetails: AddressDetails,
-  PrimaryContactDetails: Option[ContactDetails],
-  SecondaryContactDetails: Option[ContactDetails]
-)
-
-object CreateFIDetails {
-  implicit val format: OFormat[CreateFIDetails] = Json.format[CreateFIDetails]
-}
-
-case class SubmitFIDetailsResponse(fiid: Option[String])
-
-object SubmitFIDetailsResponse {
-
-  implicit val reads: Reads[SubmitFIDetailsResponse] = (json: JsValue) => {
-    val fiid = (json \ "ResponseDetails" \ "ReturnParameters" \ "Value").asOpt[String] // TODO: VALIDATE FULLY
-    JsSuccess(SubmitFIDetailsResponse(fiid))
-  }
 
 }
