+++ app/models/FinancialInstitutions/RemoveFIDetail.scala
@@ -14,19 +14,15 @@
  * limitations under the License.
  */
 
-package models.response
+package models.FinancialInstitutions
 
-import play.api.libs.json.{Json, OFormat}
+import play.api.libs.json._
 
-case class ErrorDetail(
-  timestamp: String,
-  correlationId: String,
-  errorCode: Option[String] = None,
-  errorMessage: Option[String] = None,
-  source: Option[String] = None,
-  sourceFaultDetail: Option[SourceFaultDetail] = None
+final case class RemoveFIDetail(
+  SubscriptionID: String,
+  FIID: String
 )
 
-object ErrorDetail {
-  implicit val errorDetailReads: OFormat[ErrorDetail] = Json.format[ErrorDetail]
+object RemoveFIDetail {
+  implicit val format: OFormat[RemoveFIDetail] = Json.format[RemoveFIDetail]
 }
