+++ app/models/response/SourceFaultDetail.scala
@@ -18,15 +18,12 @@ package models.response
 
 import play.api.libs.json.{Json, OFormat}
 
-case class ErrorDetail(
-  timestamp: String,
-  correlationId: String,
-  errorCode: Option[String] = None,
-  errorMessage: Option[String] = None,
-  source: Option[String] = None,
-  sourceFaultDetail: Option[SourceFaultDetail] = None
+case class SourceFaultDetail(
+  Detail: Array[String],
+  SoapFault: Option[String] = None,
+  RestFault: Option[String] = None
 )
 
-object ErrorDetail {
-  implicit val errorDetailReads: OFormat[ErrorDetail] = Json.format[ErrorDetail]
+object SourceFaultDetail {
+  implicit val sourceFaultDetailReads: OFormat[SourceFaultDetail] = Json.format[SourceFaultDetail]
 }
